requirements.txt
streamlit
pandas
numpy
yfinance
plotly
alpha_vantage
python-dotenv
